+++ src/extension.ts
@@ -1831,6 +1831,14 @@ export class ExtensionManager implements vscode.Disposable {
     }
 
     /**
+     * Appends the build directory of the active project to the current workspace
+     */
+    async appendBuildDirectoryToWorkspace()
+    {
+        await this.getActiveProject()?.appendBuildDirectoryToWorkspace();
+    }
+
+    /**
      * Show UI to allow the user to add an active configure preset
      */
     async addConfigurePreset(folder: vscode.WorkspaceFolder): Promise<boolean> {
@@ -2136,6 +2144,7 @@ async function setup(context: vscode.ExtensionContext, progress?: ProgressHandle
         'activeWorkflowPresetName',
         "useCMakePresets",
         "openCMakePresets",
+        "appendBuildDirectoryToWorkspace",
         'addConfigurePreset',
         'addBuildPreset',
         'addTestPreset',